Move configuration check to actions
